pyproject.toml
[project]
name = "rerun-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.0",
    "rerun-sdk>=0.23.3",
]

[tool.ruff.lint]
ignore = ["F541"]